Upgrade to revolution 0.3.0
The middleware stack was not clear exactly what was the list of routes
that would run. Before they were `jsx` -> `html` -> `http`, but that
made it look like the stack would run every time, but in fact, both
`jsx` and `html` were really the same thing, where `jsx` was the input
to the `html` middleware.

This flattens the entire "stack" into a single list of top level
"middlewares". If the middleware returns a JSX element, then, and only
then, will the HTML middleware transform it. But, its place in the
overall http app is the same. Because they are all in line now, it is
now clear what the precedence rules are (0 -> n) because all
middlewares are in a single list.
