+++ www/deno.json
@@ -1,6 +1,6 @@
 {
   "tasks": {
-    "dev": "deno run -A lib/watch.ts components docs lib middleware routes -- deno run -A --lock-write main.tsx"
+    "dev": "deno run -A lib/watch.ts components docs lib plugins routes -- deno run -A --lock-write main.tsx"
   },
   "lint": {
     "exclude": ["docs/esm"],
@@ -12,12 +12,13 @@
     "exclude": ["docs/esm"]
   },
   "compilerOptions": {
+    "lib": ["deno.ns", "dom.iterable", "dom"],
     "jsx": "react-jsx",
     "jsxImportSource": "revolution"
   },
   "imports": {
     "effection": "https://deno.land/x/effection@3.0.0-beta.2/mod.ts",
-    "revolution": "https://deno.land/x/revolution@0.2.3/mod.ts",
-    "revolution/jsx-runtime": "https://deno.land/x/revolution@0.2.3/jsx-runtime.ts"
+    "revolution": "https://deno.land/x/revolution@0.3.1/mod.ts",
+    "revolution/jsx-runtime": "https://deno.land/x/revolution@0.3.1/jsx-runtime.ts"
   }
 }
