+++ www/main.tsx
@@ -1,14 +1,16 @@
 import { main, suspend } from "effection";
 
-import { createRevolution, route, serveDirMiddleware } from "revolution";
+import { createRevolution, route } from "revolution";
 import { docsRoute } from "./routes/docs-route.tsx";
 import { indexRoute } from "./routes/index-route.tsx";
 import { v2docsRoute } from "./routes/v2docs-route.tsx";
-import { baseUrlMiddleware } from "./middleware/base-url.ts";
-import { rebaseMiddleware } from "./middleware/rebase.ts";
-import { etagMiddleware } from "./middleware/etag.ts";
-import { twindRevolution } from "./lib/twind.ts";
+import { assetsRoute } from "./routes/assets-route.ts";
+
 import { config } from "./twind.config.ts";
+
+import { twindPlugin } from "./plugins/twind.ts";
+import { rebasePlugin } from "./plugins/rebase.ts";
+import { etagPlugin } from "./plugins/etag.ts";
 
 import { loadDocs } from "./docs/docs.ts";
 import { loadV2Docs } from "./docs/v2-docs.ts";
@@ -21,30 +23,21 @@ await main(function* () {
 
   let docs = yield* loadDocs();
 
-  let app = createRevolution({
-    jsx: [
+  let revolution = createRevolution({
+    app: [
       route("/", indexRoute()),
       route("/docs/:id", docsRoute(docs)),
-    ],
-    html: [
-      rebaseMiddleware(),
-      twindRevolution({ config }),
-    ],
-    http: [
+      route("/assets(.*)", assetsRoute("assets")),
       route("/V2(.*)", v2docsRoute(v2docs)),
-      route(
-        "/assets(.*)",
-        serveDirMiddleware({
-          fsRoot: new URL(import.meta.resolve("./assets")).pathname,
-          urlRoot: "assets",
-        }),
-      ),
-      baseUrlMiddleware(),
-      etagMiddleware(),
+    ],
+    plugins: [
+      twindPlugin({ config }),
+      etagPlugin(),
+      rebasePlugin(),
     ],
   });
 
-  let server = yield* app.start();
+  let server = yield* revolution.start();
   console.log(`www -> http://${server.hostname}:${server.port}`);
 
   yield* suspend();
