+++ www/routes/assets-route.ts
@@ -1,0 +1,8 @@
+import { type HTTPMiddleware, serveDirMiddleware } from "revolution";
+
+export function assetsRoute(dir: string): HTTPMiddleware {
+  return serveDirMiddleware({
+    fsRoot: new URL(import.meta.resolve(`../${dir}`)).pathname,
+    urlRoot: dir,
+  });
+}
